Merged PR 54500: Add number mutation on Blur
## What's being changed

We have added additional event listeners for the Blur event (clicking away from the input in question) to the Intl input to mutate the number by prefixing the selected country code.

## Why it's being changed

We added the event listener and mutation to try to mimic the Luma behavior as closely as possible for consistency across the compat modules.

## How to review / test this change

- On any of the forms containing the Intl input field, input a valid mobile number.
- Observe the number mutate to include the selected country code.

## Notes

The mutation will not occur if the number provided is not valid.

Related work items: #255934