Merged PR 58121: Hyva checkout Phone validation tweaks
## What's being changed

We have added a guard to the phone validation function that will allow it to pass when there is no number present.

We have added toggle visibility to the subscription screen and allowed the intInput function to append additional validation rules when required.

## Why it's being changed

The requirement was interfering with allowing optional numbers to be passed.

The removal of the required attribute caused some regressions, allowing users to submit a consent form without having a phone number present.

## How to review / test this change

- Log in as a customer,
- Navigate to the subscriptions page,
- Save a valid phone number,
- Uncheck the box and save,
- Observe the phone number has been cleared.
- Disable Validation via the configuration one Website level. ( keep default enabled )
- Observe the phone validation field does not ahve Intl input on the telephone inputs across the site.

- Review test steps in [Validation and new form elements](https://dev.azure.com/dotdigital/ec/_git/dd-hyva-checkout-magento2-sms/pullrequest/54302)

Related work items: #269614, #269615, #269618