Merged PR 64023: Fix handling of additional rules in js-address.phtml
## What's being changed

The calling of `intlInput()` in js-address.phtml.

## Why it's being changed

As it was we were seeing a console error:
```
Uncaught SyntaxError: Unexpected token '&'
```

I guess this must stem from upgrading one of the `hyva-themes` packages, because we would have noticed it before.

When I looked into this I realised our logic was wrong here:
```
<?= $telephoneInputConfig->isRequired() ? $escaper->escapeHtmlAttr(json_encode(["required"])) : 'false'; ?>
```
If telephone number is not required in the customer address, we should pass an empty array for the 3rd arg and `false` as an additional 4th arg.

## How to review / test this change

- Turn on telephone number validation in Transactional SMS > Transactional SMS Settings
- In Hyva, log in as a customer
- Add an address or check a stored address
- You should see a required telephone field, with IntlTelInput rendered in the telephone input with a flag, doing validation
- Now set Customers > Customer Configuration > Name and Address Options > Show Telephone to 'Optional'
- Refresh the account screen
- You should see the flag and validation should happen, but the field is not not required, and the form can be submitted without a number

Related work items: #294627